+++ enterprise/admin/forms.py
@@ -240,14 +240,33 @@ class ManageLearnersForm(forms.Form):
         cleaned_data[self.Fields.MODE] = mode
         cleaned_data[self.Fields.NOTIFY] = self.clean_notify()
 
+        self._validate_bulk_upload()
         self._validate_course()
         self._validate_reason()
 
         return cleaned_data
 
+    def _validate_bulk_upload(self):
+        """
+        Verify that when a csv file contains a "course_id" column, there is not a singular course ID
+        specified. Also verify that a course mode is specified, which is applied to all enrollments.
+        """
+        bulk_upload_csv = self.cleaned_data.get(self.Fields.BULK_UPLOAD)
+        if bulk_upload_csv:
+            bulk_csv_contents = bulk_upload_csv.read()
+            csv_column_names = bulk_csv_contents.decode('utf-8').split('\n', 1)[0].split(',')
+            if self.CsvColumns.COURSE_ID in csv_column_names:
+                # course id column exists in csv, so validate conditionally required fields
+                if self.cleaned_data.get(self.Fields.COURSE):
+                    raise ValidationError(ValidationMessages.BOTH_COURSE_FIELDS_SPECIFIED)
+                if not self.cleaned_data.get(self.Fields.COURSE_MODE):
+                    raise ValidationError(ValidationMessages.COURSE_WITHOUT_COURSE_MODE)
+                if not self.cleaned_data.get(self.Fields.REASON):
+                    raise ValidationError(ValidationMessages.MISSING_REASON)
+
     def _validate_course(self):
         """
-        Verify that the selected mode is valid for the given course .
+        Verify that the selected mode is valid for the given course.
         """
         # Verify that the selected mode is valid for the given course .
         course_details = self.cleaned_data.get(self.Fields.COURSE)